Do not try to parse JSON if response body is empty.
Otherwise JSON.parse throws "SyntaxError: Unexpected end of JSON input" which result in ommited HTTP response headers.
Firefox and Chrome do not ommit the headers when doing the same XMLHttpRequest.
Sending an empty body with JSON as content type is not wrong in my option when doing a POST returning a 201 (e.g. Created) with a proper Location header leading to the created resource as JSON.